requirements.txt
streamlit
pandas
plotly.express 
os
numpy
matplotlib
seaborn
scikit-learn
